web_search_summerizer/pyproject.toml
[project]
name = "web_search_summerizer"
version = "0.1.0"
description = "This project summarizes web search results using AI. It fetches content from URLs, processes it, and generates concise summaries."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.0",
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.43.0",
    "lxml>=6.0.2",
    "python-decouple>=3.8",
    "readability-lxml>=0.8.4.1",
    "streamlit>=1.50.0",
]
